requirements.txt
nextcord
requests
asyncio
python-dotenv
beautifulsoup4
